requirements.txt
aiosqlite==0.17.0
discord.py==1.7.2
Unidecode==1.2.0
git+git://github.com/Rapptz/discord-ext-menus.git
git+git://github.com/Snaptraks/aiowolframalpha.git
fuzzywuzzy
numpy==1.20.3
